refactor: 현재 시간 조회 부분 분리
