+++ usecase/src/main/java/org/depromeet/spot/usecase/port/in/util/TimeUsecase.java
@@ -1,0 +1,8 @@
+package org.depromeet.spot.usecase.port.in.util;
+
+import java.time.LocalDateTime;
+
+public interface TimeUsecase {
+
+    LocalDateTime getNow();
+}
